Initial implementation of PPO Agent
This is a big patch, but this was accomplished:

1) Introducing everything required to train a PPO
Agent under models.py

The ActorCriticNN is the neural network powering
the PPO agents. Once trained, we can load
the model and inject it directly onto one of the agents
in g07agents.py

Training a PPO involves playing on the environment,
and collecting experiences that are added onto a buffer.
Every so often, we use those experiences to optimize
the policy. PPO basically "searches" onto a neighboring
space for the number of epochs passed onto the optimization.

One important thing to note is that action masking
is not yet used! We have to incorporate it, somehow.

Also, we need to work on collecting improvement
metadata to track progress.

2) The main method showcases how to train the PPO
agent.

Training is guided by a runner, which happens to be main.py